+++ src/components/popup/popup.ts
@@ -50,6 +50,8 @@ function isVirtualElement(e: unknown): e is VirtualElement {
  * @cssproperty [--auto-size-available-height] - A read-only custom property that determines the amount of height the
  *  popup can be before overflowing. Useful for positioning child elements that need to overflow. This property is only
  *  available when using `auto-size`.
+ * @cssproperty [--show-duration=100ms] - The show duration to use when applying built-in animation classes.
+ * @cssproperty [--hide-duration=100ms] - The hide duration to use when applying built-in animation classes.
  */
 @customElement('wa-popup')
 export default class WaPopup extends WebAwesomeElement {
@@ -258,7 +260,8 @@ export default class WaPopup extends WebAwesomeElement {
       this.anchorEl = this.querySelector<HTMLElement>('[slot="anchor"]');
     }
 
-    // If the anchor is a <slot>, we'll use the first assigned element as the target since slots use `display: contents`
+    // If the anchor is a `<slot>`, we'll use the first assigned element as the target since slots use
+    // `display: contents`
     // and positioning can't be calculated on them
     if (this.anchorEl instanceof HTMLSlotElement) {
       this.anchorEl = this.anchorEl.assignedElements({ flatten: true })[0] as HTMLElement;